# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Product:
    properties:
      brand:
        example: Haribo
        type: string
      id:
        example: "1"
        type: string
      name:
        example: Rotella
        type: string
      price:
        example: 13.9
        type: number
    type: object
host: localhost:8080
info:
  contact:
    name: DELGEHIER Cedric
    url: https://twitter.com/JackNemrod
  description: This is the MyApp API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MyApp API
  version: "1.0"
paths:
...
  /api/v1/products/{id}:
    get:
      consumes:
      - application/json
      description: Get a product by its ID
      parameters:
      - description: The ID of the product
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: error
          schema:
            type: string
      summary: Show product by ID
      tags:
      - products
...
swagger: "2.0"